% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreFunctions.R
\name{ClusterPrototypes}
\alias{ClusterPrototypes}
\title{Generate Self Organizing Maps
A self organizing map of the marker intensities is generated using the Yasomi package
The grid size is determined automatically}
\usage{
ClusterPrototypes(som_model, numClusters = NULL)
}
\arguments{
\item{som_model}{the self organizing map}

\item{numClusters}{the number of clusters to generate}
}
\value{
the cluster labels
}
\description{
Generate Self Organizing Maps
A self organizing map of the marker intensities is generated using the Yasomi package
The grid size is determined automatically
}
