% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreFunctions.R
\name{ComputeGridSize}
\alias{ComputeGridSize}
\title{Estimate the optimal grid size for the Self Organizing Map}
\usage{
ComputeGridSize(dataset)
}
\arguments{
\item{dataset}{The optimal grid size.}
}
\value{
the optimal grid size.
}
\description{
The function finds the eigenvalues of the sample covariance matrix.
It will then return the number of significant eigenvalues according to
the Tracy-Widom test.
The function is based on the estKW function fromt he SC3 package
}
