% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilityFunctions.R
\name{.arsinhNnorm}
\alias{.arsinhNnorm}
\title{Function to do arsinh normalization}
\usage{
.arsinhNnorm(x, cofactor = 5)
}
\arguments{
\item{x}{A numeric or complex vector}

\item{cofactor}{Cofactor of the vector. Default is 5.}
}
\value{
Arsinh normalized vector.
}
\description{
Function to do arsinh normalization
}
