% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilityFunctions.R
\name{.computeElbow}
\alias{.computeElbow}
\title{A function to compute the elbow point given a set of points}
\usage{
.computeElbow(vals)
}
\arguments{
\item{vals}{Values to compute the elbow point of.}
}
\value{
A integer indicating the elbow point of \code{vals}.
}
\description{
A function to compute the elbow point given a set of points
}
