% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilityFunctions.R
\name{.minmaxNorm}
\alias{.minmaxNorm}
\title{Function to do min max normalization}
\usage{
.minmaxNorm(x)
}
\arguments{
\item{x}{Matrix to min max nomalize.}
}
\description{
Function to do min max normalization
}
