% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilityFunctions.R
\name{.uniformData}
\alias{.uniformData}
\title{Creates uniformly distributed data of same dimensionality as input data
this function was obtained from the Stab package}
\usage{
.uniformData(data)
}
\description{
Creates uniformly distributed data of same dimensionality as input data
this function was obtained from the Stab package
}
