% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreFunctions.R
\name{estimateNumcluster}
\alias{estimateNumcluster}
\title{A function for estimating the number of clusters using various method
Methods available are: Discriminant, Distance
(Gap, Silhouette, Slope, Jump, and Within Cluster Distance,) and Instability}
\usage{
estimateNumcluster(
  som_model,
  method = c("Discriminant", "Distance", "Stability"),
  kseq = 2:20
)
}
\arguments{
\item{som_model}{the Self Organizing Map object generated by generatePrototypes()}

\item{method}{one of Discriminant, Distance, Stability. By default, everything is run}

\item{kseq}{a sequence of the number of clusters to try. Default is 2:20 clusters}
}
\value{
A list containing the optimal number of clusters for each class of method and the
values over the sequence of k provided
}
\description{
A function for estimating the number of clusters using various method
Methods available are: Discriminant, Distance
(Gap, Silhouette, Slope, Jump, and Within Cluster Distance,) and Instability
}
\author{
Elijah WIllie \href{mailto:ewil3501@uni.sydney.edu.au}{ewil3501@uni.sydney.edu.au}
}
