% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FuseSOM.R
\docType{package}
\name{FuseSOM}
\alias{FuseSOM}
\title{FuseSOM: FuseSOM provides a pipeline for the clustering of highly
multiplexed in situ imaging cytometry assays. This pipeline uses
the Self Organizing Map architecture coupled with Multiview hierarchical
clustering. We also provide functions for normalisation and estimation
of the number of clusters.}
\description{
The FuseSOM package provides three categories of important functions:
foo, bar and baz.
}
