% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreFunctions.R
\name{generateHeatmap}
\alias{generateHeatmap}
\title{Generate Heatmap of Clusters
A heatmap of marker intensities given cluster labels is generated}
\usage{
generateHeatmap(data, markers, labels)
}
\arguments{
\item{data}{the marker intensities}

\item{markers}{the markers of interest}

\item{heatmap}{the heatmap of clusters by markers}
}
\value{
the cluster label
}
\description{
Generate Heatmap of Clusters
A heatmap of marker intensities given cluster labels is generated
}
