% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreFunctions.R
\name{GeneratePrototypes}
\alias{GeneratePrototypes}
\title{Generate Self Organizing Maps
A self organizing map of the marker intensities is generated using the Yasomi package
The grid size is determined automatically}
\usage{
GeneratePrototypes(data)
}
\arguments{
\item{data}{the marker intensities}
}
\value{
the self organizing map.
}
\description{
Generate Self Organizing Maps
A self organizing map of the marker intensities is generated using the Yasomi package
The grid size is determined automatically
}
