% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreFunctions.R
\name{markerHeatmap}
\alias{markerHeatmap}
\title{A function for generating a heat map of marker expression across clusters}
\usage{
markerHeatmap(data, markers = NULL, clusters = NULL)
}
\arguments{
\item{data}{a matrix or dataframe where the rows are samples and columns are
markers}

\item{markers}{a list of markers of interest. If not provided, all columns
will be used}

\item{clusters}{a vector of cluster labels}
}
\value{
an elbow plot object where the optimal number of clusters is marked
}
\description{
A function for generating a heat map of marker expression across clusters
}
\author{
Elijah WIllie \href{mailto:ewil3501@uni.sydney.edu.au}{ewil3501@uni.sydney.edu.au}
}
