% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreFunctions.R
\name{NormalizeData}
\alias{NormalizeData}
\title{Normalize Marker Intensities
The matrix of intensities is normalized based on one of four different method
These methods include Percentile, zscore, arsinh and minmax}
\usage{
NormalizeData(data, markers, method = "none")
}
\arguments{
\item{data}{the raw intensity scores.}

\item{markers}{the markers of interest.}

\item{method}{the normalizaton method}
}
\value{
normalized matrix.
}
\description{
Normalize Marker Intensities
The matrix of intensities is normalized based on one of four different method
These methods include Percentile, zscore, arsinh and minmax
}
