% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreFunctions.R
\name{plotSOM}
\alias{plotSOM}
\title{A function for plotting the self organizing map
function was obtained from https://rdrr.io/rforge/yasomi/ with some minor modifications}
\usage{
plotSOM(som, border = NA, withGrid = TRUE, ...)
}
\arguments{
\item{som}{the self organizing map}

\item{border}{a value for the plot border}

\item{withGrid}{option to add the som grid}
}
\value{
None
}
\description{
A function for plotting the self organizing map
function was obtained from https://rdrr.io/rforge/yasomi/ with some minor modifications
}
\examples{
data("risom_dat")
risomMarkers <- c('CD45','SMA','CK7','CK5','VIM','CD31','PanKRT','ECAD')
prototypes <- generatePrototypes(risom_dat[, risomMarkers])
plotSOM(prototypes)

}
\author{
Elijah WIllie \href{mailto:ewil3501@uni.sydney.edu.au}{ewil3501@uni.sydney.edu.au}
}
