% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risom-data.R
\docType{data}
\name{risom_dat}
\alias{risom_dat}
\title{IMC Breast Cancer Data
Data from A spatial atlas of breast cancer progression using MIBI-TOF
and tissue transcriptomics}
\format{
An object of class \code{"data.frame"}.
}
\source{
Mendeley Data, \url{https://data.mendeley.com/datasets/d87vg86zd8/3}
}
\usage{
data(risom)
}
\description{
IMC Breast Cancer Data
Data from A spatial atlas of breast cancer progression using MIBI-TOF
and tissue transcriptomics
}
\references{
T. Risom, et al.
Transition to invasive breast cancer is associated with progressive changes
in the structure and composition of tumor stroma Cell, 185 (2022), pp. 299-310
(\href{https://www.sciencedirect.com/science/article/pii/S0092867421014860?via\%3Dihub}{ScienceDirect})
}
\keyword{datasets}
