% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreFunctions.R
\name{runFuseSOM}
\alias{runFuseSOM}
\title{A wrapper function to run the FuseSOM algorithm.
This function accepts a matrix, dataframe or a SingleCellExperiment object}
\usage{
runFuseSOM(data, assay = NULL, markers = NULL, numClusters = NULL)
}
\arguments{
\item{data}{a matrix, dataframe or SingleCellExperiment object. For matrices
and dataframes, it is assumed that markers are the columns and samples rows}

\item{assay}{the assay of interest if SingleCellExperiment object is used}

\item{markers}{the markers of interest. If this is not provided, all columns will be used}

\item{numCluster}{the number of clusters to be generated from the data}
}
\value{
A list conataining the SOM model and the cluster labels if a dataframe
or matrix is provided

A Slist containing the SOM model and a ingleCellExperiment object with
labels in coldata if a SingleCellExperiment object is provided
}
\description{
A wrapper function to run the FuseSOM algorithm.
This function accepts a matrix, dataframe or a SingleCellExperiment object
}
\author{
Elijah WIllie \href{mailto:ewil3501@uni.sydney.edu.au}{ewil3501@uni.sydney.edu.au}
}
