% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/som.R
\name{somInitPca.default}
\alias{somInitPca.default}
\title{these functions were obtained from https://rdrr.io/rforge/yasomi/ with some major modifications}
\usage{
\method{somInitPca}{default}(data, somGrid, weights, with.princomp = FALSE, ...)
}
\description{
these functions were obtained from https://rdrr.io/rforge/yasomi/ with some major modifications
}
