% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreFunctions.R
\name{clusterPrototypes}
\alias{clusterPrototypes}
\title{Cluster the prototypes from the Self Organizing Map}
\usage{
clusterPrototypes(som_model, numClusters = NULL)
}
\arguments{
\item{som_model}{the self organizing map}

\item{numClusters}{the number of clusters to generate}
}
\value{
the cluster labels
}
\description{
Clustering is done using hierarchical clustering with
the average linkage function
}
\examples{
 
data("risom_dat")
risom_markers <- c('CD45','SMA','CK7','CK5','VIM','CD31','PanKRT','ECAD')
prototypes <- generatePrototypes(risom_dat[, risom_markers])
clusters <- clusterPrototypes(prototypes, 23)

}
