% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreFunctions.R
\name{ClusterPrototypes}
\alias{ClusterPrototypes}
\title{Generate Self Organizing Maps
The prototypes generated from the self organizing map is clustered using
hierarchical clustering based on the multiview integration of the Pearson,
Cosine and Spearman distances between the protoypes}
\usage{
ClusterPrototypes(som_model, numClusters = NULL)
}
\arguments{
\item{som_model}{the self organizing map}

\item{numClusters}{the number of clusters to generate}
}
\value{
the cluster labels
}
\description{
Generate Self Organizing Maps
The prototypes generated from the self organizing map is clustered using
hierarchical clustering based on the multiview integration of the Pearson,
Cosine and Spearman distances between the protoypes
}
