% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreFunctions.R
\name{computeGridsize}
\alias{computeGridsize}
\title{Estimate the optimal grid size for the Self Organizing Map}
\usage{
computeGridsize(dataset)
}
\arguments{
\item{dataset}{The optimal grid size.}
}
\value{
the optimal grid size.
}
\description{
The function finds the eigenvalues of the sample covariance matrix.
It will then return the number of significant eigenvalues according to
the Tracy-Widom test.
The function is based on the estKW function from the SC3 package
}
\author{
Elijah WIllie \href{mailto:ewil3501@uni.sydney.edu.au}{ewil3501@uni.sydney.edu.au}
}
