% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilityFunctions.R
\name{.computeElbow}
\alias{.computeElbow}
\title{A function to compute the elbow point given a set of points}
\usage{
.computeElbow(vals)
}
\description{
A function to compute the elbow point given a set of points
}
