% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilityFunctions.R
\name{.percentileNorm}
\alias{.percentileNorm}
\title{Function to do percentile normalizaton}
\usage{
.percentileNorm(x)
}
\arguments{
\item{x}{Matrix to percentile normilse.}
}
\description{
Function to do percentile normalizaton
}
