% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runDiscriminant.R
\name{.runDiscriminant}
\alias{.runDiscriminant}
\title{Discriminant cluster estimator}
\usage{
.runDiscriminant(distMat, minClusterSize, alpha = 0.001)
}
\arguments{
\item{distMat}{A distance matrix}

\item{minClusterSize}{The minimum cluster size}

\item{alpha}{a value between 0 and 1 specifying the desired level of cutoff}
}
\value{
Optimal number of clusters
}
\description{
Function to estimate the number of clusters using discriminant analysis
parts of this function is based on the sigclust2 package by Patrick Kimes
see https://github.com/pkimes/sigclust2
}
