% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreFunctions.R
\name{estimateNumCluster}
\alias{estimateNumCluster}
\title{A function for estimating the number of clusters using various method
Methods available are: Discriminant, Distance
(Gap, Silhouette, Slope, Jump, and Within Cluster Distance,) and Instability}
\usage{
estimateNumCluster(
  data,
  method = c("Discriminant", "Distance", "Stability"),
  kseq = 2:20
)
}
\arguments{
\item{data}{the Self Organizing Map object generated by generatePrototypes(), or
an object of class SingleCellExperiment or SpatialExperiment}

\item{method}{one of Discriminant, Distance, Stability. By default, everything is run}

\item{kseq}{a sequence of the number of clusters to try. Default is 2:20 clusters}
}
\value{
A list containing the cluster estimations if a dataframe
or matrix is provided

A SingleCellExperiment object with the cluster estimation in the metadata
if a SingleCellExperiment or SpatialExperiment object is provided
}
\description{
A function for estimating the number of clusters using various method
Methods available are: Discriminant, Distance
(Gap, Silhouette, Slope, Jump, and Within Cluster Distance,) and Instability
}
\author{
Elijah WIllie \href{mailto:ewil3501@uni.sydney.edu.au}{ewil3501@uni.sydney.edu.au}
}
