% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreFunctions.R
\name{generatePrototypes}
\alias{generatePrototypes}
\title{Generate a Self Organizing Map and return it's prototypes
A self organizing map of the marker intensities is generated using the Yasomi package
The grid size is determined automatically}
\usage{
generatePrototypes(data, verbose = FALSE)
}
\arguments{
\item{data}{the marker intensities}
}
\value{
the self organizing map object.
}
\description{
Generate a Self Organizing Map and return it's prototypes
A self organizing map of the marker intensities is generated using the Yasomi package
The grid size is determined automatically
}
