% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreFunctions.R
\name{optiPlot}
\alias{optiPlot}
\title{A function generating the elbow plot for the optimal number of clusters
returned by the estimateNumcluster() function
Methods available are:
Gap, Silhouette, Slope, Jump, and Within Cluster Distance(WCD)}
\usage{
optiPlot(k_est, method = "jump")
}
\arguments{
\item{k_est}{the estimation list returned by estimateNumcluster()}

\item{method}{one of 'jump', 'slope', 'wcd', 'gap', or 'silhouette'}
}
\value{
an elbow plot object where the optimal number of clusters is marked
}
\description{
A function generating the elbow plot for the optimal number of clusters
returned by the estimateNumcluster() function
Methods available are:
Gap, Silhouette, Slope, Jump, and Within Cluster Distance(WCD)
}
\author{
Elijah WIllie \href{mailto:ewil3501@uni.sydney.edu.au}{ewil3501@uni.sydney.edu.au}
}
