% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreFunctions.R
\name{runFuseSOM}
\alias{runFuseSOM}
\title{A wrapper function to run the FuseSOM algorithm.
This function accepts a matrix, dataframe or a SingleCellExperiment object}
\usage{
runFuseSOM(
  data,
  markers = NULL,
  numClusters = NULL,
  assay = NULL,
  clusterCol = "clusters",
  verbose = FALSE
)
}
\arguments{
\item{data}{a matrix, dataframe, SingleCellExperiment or SpatialExperiment object.
For matrices and dataframes, it is assumed that markers are the columns and samples rows}

\item{markers}{the markers of interest. If this is not provided, all columns will be used}

\item{assay}{the assay of interest if SingleCellExperiment object is used}

\item{clusterCol}{the name of the column to store the clusters in}

\item{verbose}{should the generation of the Self Organising Map be printed}

\item{numCluster}{the number of clusters to be generated from the data}
}
\value{
A list containing the SOM model and the cluster labels if a dataframe
or matrix is provided

A SingleCellExperiment object with labels in coldata, and the SOM model
in metadata if a SingleCellExperiment or SpatialExperiment object is provided
}
\description{
A wrapper function to run the FuseSOM algorithm.
This function accepts a matrix, dataframe or a SingleCellExperiment object
}
\author{
Elijah WIllie \href{mailto:ewil3501@uni.sydney.edu.au}{ewil3501@uni.sydney.edu.au}
}
